Add files via upload
These MATLAB scripts implement the models used in "Kinetic and mechanistic study of CO2 absorption into vanadium-promoted aqueous K2CO3". The repository includes "Example.m", which demonstrates how to run the code and produce the results. The accompanying functions support the main script and are not intended to be run independently. To reproduce the results, run Example.m after ensuring all files are in the same folder or added to the MATLAB path.